Emulate the `getrandom` syscall.
Calls to the Linux `getrandom` syscall can block during early boot because
there is not a lot of entropy available. We don't really need entropy for
anything, but dependencies use `HashMap` which requires entropy to
initialize. This module installs a seccomp filter that catches `getrandom`
syscalls and emulates the call using the `RDRAND` instruction.